Extract LLVM-related functions to `tools`
This will help to improve the testability of our code that calls
directly into LLVM, and also help reuse that code to improve the rest of
our testing framework.
