+++ build.sbt
@@ -228,11 +228,13 @@ lazy val nativelib =
       val clang   = nativeClang.value
       val clangpp = nativeClangPP.value
       val source  = baseDirectory.value
+      val logger = sys.process.ProcessLogger(l => streams.value.log.info(l),
+                                             l => streams.value.log.error(l))
       val compileSuccess =
         IO.withTemporaryDirectory { tmp =>
           IO.copyDirectory(baseDirectory.value, tmp)
-          scala.scalanative.sbtplugin.ScalaNativePluginInternal
-            .compileCSources(clang, clangpp, tmp, streams.value.log)
+          scala.scalanative.llvm.LLVM
+            .compileCSources(clang, clangpp, tmp, logger)
         }
       if (compileSuccess) {
         (compile in Compile).value
